Remove Repository interface * The abstraction is  at the wrong level, causing more overhead than it helps.
